update iPEST
git-svn-id: https://svn.oss.deltares.nl/repos/imod/trunk@2887 8a8853d3-af7d-47ec-b359-4bd2096ba547

+++ iMOD/SRC/imod_residualplot/imod_residualplot_par.f90
@@ -29,10 +29,10 @@ TYPE IPFRESOBJ
  REAL(KIND=DP_KIND),POINTER,DIMENSION(:) :: M=>NULL()              !## computed values
  REAL(KIND=DP_KIND),POINTER,DIMENSION(:) :: O=>NULL()              !## observation
  REAL(KIND=DP_KIND),POINTER,DIMENSION(:) :: W=>NULL()              !## weight value
- REAL(KIND=DP_KIND),POINTER,DIMENSION(:) :: NS=>NULL()              !## weight value
- CHARACTER(LEN=32),POINTER,DIMENSION(:) :: C=>NULL()               !## name
- INTEGER,POINTER,DIMENSION(:) ::            L=>NULL()              !## layer
- INTEGER(KIND=DP_KIND),POINTER,DIMENSION(:) :: D=>NULL()           !## date
+ REAL(KIND=DP_KIND),POINTER,DIMENSION(:) :: NS=>NULL()             !## weight value
+ CHARACTER(LEN=32),POINTER,DIMENSION(:) ::  C=>NULL()              !## name
+ INTEGER,POINTER,DIMENSION(:) :: L=>NULL()                         !## layer
+ INTEGER(KIND=8),POINTER,DIMENSION(:) :: D=>NULL()                 !## date
  INTEGER :: NPOINTS                                    !## number of points in ipf
 END TYPE IPFRESOBJ
 TYPE(IPFRESOBJ),DIMENSION(:),ALLOCATABLE :: IPFR       !## stores original values from txtfile per ipf
